requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cloudinary==1.38.0
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
